requirements.txt
PyQt6
pyqt6-tools
pyqt6-webengine
pynput
requests
qtawesome
psutil
pywin32
win32gui
win32con
win32api